requirements.txt
streamlit
openai
pinecone-client
python-dotenv
langchain
PyMuPDF
tiktoken
